functions/image_processor/requirements.txt
google-cloud-storage>=2.0.0
Pillow>=9.0.0